Add files via upload
This is what I have so far.
I basically created a few windows using QT with c++ code. When a user login with admin account, the program will open a window for admin, and I plan to add functions like editing and deleting to this window. Also, a user can register an account, and if the user use that account to login, he/she can only view the information for this account.